// src/types/enums/wallet.ts
export enum WalletID {
  Cosmostation = 'cosmostation',
  CosmostationMobile = 'mobile-cosmostation',
  Keplr = 'keplr',
  KeplrMobile = 'mobile-keplr',
  Leap = 'leap-cosmos',
  LeapMobile = 'mobile-leap-cosmos',
  LeapSnap = 'leap-metamask-cosmos-snap',
  Station = 'station',
  Vectis = 'vectis-cosmos',
  Xdefi = 'xfi-cosmos',
}

export enum ChainInfoID {
  Osmosis1 = 'osmosis-1',
  OsmosisDevnet = 'devnet',
  Neutron1 = 'neutron-1',
  NeutronTestnet = 'pion-1',
}